Remove and redistribute contents of rastervision_examples package
* Examples should live inside the packages they pertain to
* cookiecutter_template isn't valid Python code (it's a templatized repo)
and shouldn't be inside a Python package, so is moved to the top-level
