+++ cookiecutter_template/{{cookiecutter.project_name}}/Dockerfile
