+++ cookiecutter_template/{{cookiecutter.project_name}}/configs/test.py
