+++ cookiecutter_template/{{cookiecutter.project_name}}/docker/ecr_publish
