+++ cookiecutter_template/{{cookiecutter.project_name}}/docker/run
