+++ cookiecutter_template/{{cookiecutter.project_name}}/rastervision2/{{cookiecutter.project_name}}/__init__.py
