+++ cookiecutter_template/{{cookiecutter.project_name}}/rastervision2/{{cookiecutter.project_name}}/test_pipeline_config.py
