+++ cookiecutter_template/{{cookiecutter.project_name}}/scripts/debug
