--- rastervision_examples/rastervision/examples/semantic_segmentation/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-
-import rastervision.pipeline
