--- rastervision_examples/requirements.txt
@@ -1,1 +1,0 @@
-rastervision_pytorch_backend==0.12.*
