+++ rastervision_pipeline/rastervision/examples/pipeline_example1/__init__.py
