+++ rastervision_pipeline/rastervision/examples/pipeline_example1/my_config.py
