+++ rastervision_pipeline/rastervision/examples/pipeline_example1/sample_pipeline.py
