+++ rastervision_pipeline/rastervision/examples/pipeline_example2/__init__.py
