+++ rastervision_pipeline/rastervision/examples/pipeline_example2/my_config.py
