+++ rastervision_pipeline/rastervision/examples/pipeline_example2/sample_pipeline2.py
