+++ rastervision_pipeline/rastervision/examples/pipeline_example3/__init__.py
