+++ rastervision_pipeline/rastervision/examples/pipeline_example3/deluxe_message_maker.py
