+++ rastervision_pipeline/rastervision/examples/pipeline_example3/my_config.py
