+++ rastervision_pytorch_backend/rastervision/pytorch_backend/examples/chip_classification/__init__.py
