+++ rastervision_pytorch_backend/rastervision/pytorch_backend/examples/semantic_segmentation/isprs_potsdam.py
@@ -9,7 +9,7 @@ from rastervision.core.data import *
 from rastervision.core.analyzer import *
 from rastervision.pytorch_backend import *
 from rastervision.pytorch_learner import *
-from rastervision.examples.utils import get_scene_info, save_image_crop
+from rastervision.pytorch_backend.examples.utils import get_scene_info, save_image_crop
 
 
 def get_config(runner, raw_uri, processed_uri, root_uri, test=False):
