+++ rastervision_pytorch_backend/requirements.txt
@@ -1,3 +1,4 @@
 rastervision_pipeline==0.12.*
 rastervision_core==0.12.*
 rastervision_pytorch_learner==0.12.*
+rastervision_aws_s3==0.12.*
